Test work
createTest will now simply generate a test and redirect you to the list
of tests created, I have put in a timestamp so that the instructor can
find the one that they have created. Implemented a UUID so that the ID
is automatically created for each test.
Test/id will now take you to the test and load the questions.
STILL NEED TO HANDLE ANSWERING THE QUESTIONS.